requirements.txt
google-api-python-client
google-auth-oauthlib
openpyxl
PyDrive2
selenium
fpdf
Flask==2.2.2
mysql-connector-python
python-dotenv
cryptography
pygps==1.3.5
